requirements.txt
Pillow
requests
opencv-python
pyinstaller